pyproject.toml
[tool.poetry]
name = "fitnessbot"
version = "0.1.0"
description = ""
authors = ["Vladyslav Ushakov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
python-telegram-bot = "^20.4"
sqlalchemy = "^2.0.20"
filters = "^1.3.2"
python-decouple = "^3.8"
psycopg2 = "^2.9.9"
alembic = "^1.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
